Abstract of the project
According to the author (Keating, 20147), Cervical cancer is the second most hazardous and deadly disease after Breast cancer known to mankind. Over 250 women in Ireland are found to have cervical cancer every year. This count is increasing day by day. The reason researchers finding a way to detect early is , at later stages of cancer it is highly risky to save the patient from cancer as it spreads rapidly once detected in cervix area. Now the recent study states that there is research going for cervical cancer detection using machine learning algorithm. This study uses the screening images of Pap test and performs classification on them to see the cause of cancer. On the other hand, many instruments are created which helps in testing a patient’s cells to detect any type of disease not only cancer. Cervical cancer which was difficult or rather impossible to cure in past is now cured using the new developments in technology. Machine learning is a powerful technology used in all the fields like weather storm detection, Banking, Businesses, Medical fields etc. This is bringing a new future for everyone and researchers are studying more using machine learning to use it as a precautious weapon for all of us. Machine learning is used by scientists, biologists etc to know the future events in advance like a forecast. Every day new study is found by a team which brings a lot of helpful knowledge and it has cured thousands of lakhs of patients.